refactor(mcp): split manager.rs monolith into focused submodules
Split crates/zeph-mcp/src/manager.rs (3883 lines, one monolithic
impl McpManager plus scattered free functions) into a directory module
grouped by concern:

- manager/mod.rs    shared data types, McpManager struct, McpTrustLevel
                    re-export, module declarations
- manager/builder.rs   construction and fluent configuration builders
- manager/connect.rs    connect_all/OAuth flows, probing, refresh task
- manager/retry.rs      connection retry, backoff, per-server connect
- manager/server.rs     dynamic add/remove, queries, instructions, shutdown
- manager/call.rs       call_tool entry point and result extraction
- manager/ingest.rs     sanitize/attest/allowlist pipeline, trust penalties
- manager/tests.rs      relocated unit tests

Public API is unchanged: McpManager, McpTransport, ServerConnectOutcome,
and ServerEntry remain re-exported from the crate root, all inherent
methods keep their original visibility, and crate::manager::McpTrustLevel
still resolves for policy.rs. Cross-submodule helpers are bumped to
pub(super); no behavior change.

Closes #4919
